refractor(menuScripts) : reorganize and simplify
