+++ Assets/Scripts/Menu/Selection_P2.cs
@@ -14,7 +14,7 @@ public class Selection_P2 : MonoBehaviour {
 	private Renderer render;
 	private AudioSource aud;
 
-	// Use this for initialization
+
 	void Start () {
 		aud = GetComponent<AudioSource>();
 		pos = 2;
@@ -22,7 +22,7 @@ public class Selection_P2 : MonoBehaviour {
 		render = GetComponent<Renderer>();
 	}
 
-	// Update is called once per frame
+
 	void Update () {
 		if (Input.GetAxis("J2_Deplacement") == 1){
 			if ((select_perso < 5) && (Camera_Menu.pos == pos)) {
@@ -38,7 +38,7 @@ public class Selection_P2 : MonoBehaviour {
 		}
 	}
 
-	//Autres fonctions
+
 	void Affichage (){
 		aud.PlayOneShot (son_selection);
 		switch (select_perso) {
